refactor: move routes into separate routers
